+++ src/main/java/ru/booksapi/interfaces/BookService.java
@@ -16,9 +16,9 @@ import java.util.List;
  */
 public interface BookService {
 
-    public Map<Integer,Map<String,String>> getBookById(Long id) throws ServiceExeption;
-    public Map<Integer,Map<String,String>> getBooksByAuthorId(Long authorId) throws ServiceExeption;
-    public Map<Integer,Map<String,String>> getBooksByGenreId(Long genreId) throws ServiceExeption;
+    public Map<Integer,Map<String,String>> getBookById(Integer id) throws ServiceExeption;
+    public Map<Integer,Map<String,String>> getBooksByAuthorId(Integer authorId) throws ServiceExeption;
+    public Map<Integer,Map<String,String>> getBooksByGenreId(Integer genreId) throws ServiceExeption;
     public Map<Integer,Map<String,String>> getAllBooks() throws ServiceExeption;
     public void updateBookById(Map<String,String> map) throws ServiceExeption;
     public void putNewBook(Map<String,String> map)  throws ServiceExeption;
